+++ src/LosRobles/Base.php
@@ -3,68 +3,32 @@
 namespace Fragen\LosRobles;
 
 class Base {
-
-	static $depts;
-	protected static $object = false;
+	public static $depts;
+	private static $object = false;
 
 	public static function instance() {
-		$class = __CLASS__;
 		if ( false === self::$object ) {
-			self::$object = new $class();
+			self::$object = new self();
 		}
+
 		return self::$object;
 	}
 
 	public function __construct() {
-
-		self::$depts = array(
-			'--' => '',
-			'Emergency Medicine'    => 'emergency-medicine',
-			'Family Medicine'       => 'family-medicine',
-			'Medicine'              => 'medicine',
-			'Pediatrics'            => 'pediatrics',
-			'Obstetrics/Gynecology' => 'obstetrics-gynecology',
-			'Radiology'             => 'radiology',
-			'Surgery'               => 'surgery'
-			);
-
-		add_filter( 'login_redirect', array( $this, 'change_login_redirect' ), 10, 3);
-		//add_action( 'init', array( $this, 'add_custom_taxonomies' ), 0 );
-		add_action( 'init', array( $this, 'create_post_type' ) );
-		add_action( 'plugins_loaded', array( $this, 'hide_toolbar' ) );
-
-		is_admin() ? new Admin( $this ) : new Frontend( $this );
 	}
 
-	public static function make_dropdown( $user ) {
-		$value = get_user_meta( $user->ID, 'drmc_department' );
-		if ( ! $_POST ) {
-			$_POST['drmc_department'] = array( 0 => '' );
+	public function load_hooks() {
+		add_filter( 'login_redirect', [ $this, 'change_login_redirect' ], 10, 3 );
+		// add_action( 'init', array( $this, 'add_custom_taxonomies' ), 0 );
+		add_action( 'init', [ $this, 'create_post_type' ] );
+		add_action( 'plugins_loaded', [ $this, 'hide_toolbar' ] );
+
+		if ( is_admin() ) {
+			( new Admin( $this ) )->load_hooks();
 		}
-		if ( ! $value ) {
-			$value = $_POST['drmc_department'];
-		}
-		$dropdown   = array();
-		$dropdown[] = '<select name="drmc_department" id="drmc_department">';
-		foreach ( self::$depts as $dept => $tax ) {
-			$dropdown[] = "<option value='$tax' " . selected($value[0], $tax, false) . ">$dept</option>";
-			//$dropdown[] = '<option value="'. $tax . '" <' . '?php selected( $value, "' . $tax . '" ); ?' . '>>' . $dept . '</option>';
-		}
-		$dropdown[] = '</select>';
-		$content    = implode( "\n", $dropdown );
-		echo $content;
 	}
 
-
-	public static function get_department() {
-		global $current_user;
-		get_currentuserinfo();
-		$user_dept = get_user_meta( $current_user->ID, 'drmc_department' );
-
-		return $user_dept;
-	}
-
-	//http://nathany.com/redirecting-wordpress-subscribers
+	// http://nathany.com/redirecting-wordpress-subscribers
 	public function change_login_redirect( $redirect_to, $request_redirect_to, $user ) {
 		if ( $user instanceof \WP_User && false === $user->has_cap( 'add_users' ) ) {
 			return get_bloginfo( 'siteurl' );
@@ -73,17 +37,16 @@ class Base {
 		return $redirect_to;
 	}
 
-	//http://digwp.com/2011/04/admin-bar-tricks/
+	// http://digwp.com/2011/04/admin-bar-tricks/
 	public function hide_toolbar() {
 		// show admin bar only for admins
 		if ( ! current_user_can( 'manage_options' ) ) {
-			//add_filter( 'show_admin_bar', '__return_false' );
+			// add_filter( 'show_admin_bar', '__return_false' );
 		}
 
 		// show admin bar only for admins and editors
-		//if( ! current_user_can( 'edit_posts' ) ) { add_filter( 'show_admin_bar', '__return_false' ); }
+		// if( ! current_user_can( 'edit_posts' ) ) { add_filter( 'show_admin_bar', '__return_false' ); }
 	}
-
 
 	/**
 	 * Add custom taxonomies
@@ -94,79 +57,52 @@ class Base {
 	 */
 	public function add_custom_taxonomies() {
 		// Add new "Departments" taxonomy to Posts
-		register_taxonomy( 'department', 'lrhoa_voting', array(
-			// Hierarchical taxonomy (like categories)
-			'hierarchical' => false,
-			// This array of options controls the labels displayed in the WordPress Admin UI
-			'labels'       => array(
-				'name'              => _x( 'Departments', 'taxonomy general name' ),
-				'singular_name'     => _x( 'Department', 'taxonomy singular name' ),
-				'search_items'      => __( 'Search Departments' ),
-				'all_items'         => __( 'All Departments' ),
-				'parent_item'       => __( 'Parent Department' ),
-				'parent_item_colon' => __( 'Parent Department:' ),
-				'edit_item'         => __( 'Edit Department' ),
-				'update_item'       => __( 'Update Department' ),
-				'add_new_item'      => __( 'Add New Department' ),
-				'new_item_name'     => __( 'New Department Name' ),
-				'menu_name'         => __( 'Departments' ),
-			),
-			'query_var'    => 'department',
-			// Control the slugs used for this taxonomy
-			'rewrite'      => array(
-				'slug'         => 'department', // This controls the base slug that will display before each term
-				'with_front'   => false, // Don't display the category base before "/locations/"
-				'hierarchical' => true // This will allow URL's like "/locations/boston/cambridge/"
-			),
-		));
+		register_taxonomy(
+			'department',
+			'lrhoa_voting',
+			[
+				// Hierarchical taxonomy (like categories)
+				'hierarchical' => false,
+				// This array of options controls the labels displayed in the WordPress Admin UI
+				'labels'       => [
+					'name'              => _x( 'Departments', 'taxonomy general name' ),
+					'singular_name'     => _x( 'Department', 'taxonomy singular name' ),
+					'search_items'      => __( 'Search Departments' ),
+					'all_items'         => __( 'All Departments' ),
+					'parent_item'       => __( 'Parent Department' ),
+					'parent_item_colon' => __( 'Parent Department:' ),
+					'edit_item'         => __( 'Edit Department' ),
+					'update_item'       => __( 'Update Department' ),
+					'add_new_item'      => __( 'Add New Department' ),
+					'new_item_name'     => __( 'New Department Name' ),
+					'menu_name'         => __( 'Departments' ),
+				],
+				'query_var'    => 'department',
+				// Control the slugs used for this taxonomy
+				'rewrite'      => [
+					'slug'         => 'department', // This controls the base slug that will display before each term
+					'with_front'   => false, // Don't display the category base before "/locations/"
+					'hierarchical' => true, // This will allow URL's like "/locations/boston/cambridge/"
+				],
+			]
+		);
 	}
 
 	public function create_post_type() {
-		register_post_type( 'lrhoa_voting',
-			array(
-				'labels' => array(
+		register_post_type(
+			'lrhoa_voting',
+			[
+				'labels'        => [
 					'name'          => __( 'Elections' ),
-					'singular_name' => __( 'Election' )
-				),
-			'public'        => true,
-			'menu_position' => 5,
-			'menu_icon'     => 'dashicons-chart-bar',
-			'rewrite'       => array( 'slug' => 'elections' ),
-			'taxonomies'    => array( 'department' ),
-			'supports'      => array( 'title', 'editor', 'comments', 'post-formats' )
-			)
+					'singular_name' => __( 'Election' ),
+				],
+				'public'        => true,
+				'menu_position' => 5,
+				'menu_icon'     => 'dashicons-chart-bar',
+				'rewrite'       => [ 'slug' => 'elections' ],
+				// 'taxonomies'    => [ 'department' ],
+				'supports'      => [ 'title', 'editor', 'comments', 'post-formats' ],
+			]
 		);
 	}
-
-	public function add_user_roles() {
-		remove_role( 'members' );
-		remove_role( 'non_members' );
-		add_role( 'members', 'Members',
-			array(
-					'read'                 => true,
-					'can_vote'             => true,
-					'email_single_user'    => true,
-					'email_multiple_users' => true,
-					'email_user_groups'    => true,
-			)
-		);
-		add_role( 'non_members', 'Non-Members',
-			array(
-					'read'     => true,
-					'can_vote' => false,
-			)
-		);
-	}
-
-	public function add_admin_voting() {
-		$role = get_role( 'administrator' );
-		$role->add_cap( 'can_vote' );
-		$role->add_cap( 'members' );
-	}
-
-	public static function activate() {
-		self::add_user_roles();
-		self::add_admin_voting();
-	}
-
 }