+++ CHANGES.md
@@ -1,3 +1,7 @@
+#### 1.0.0
+* refactor and general cleanup
+* add custom user fields for street number, phone number, and emergency contact
+
 #### 0.1.2
 * use Force Strong Password plugin, for all users